Sprint 3

+++ flo2d/gui/dlg_sampling_elev.py
@@ -12,16 +12,19 @@ import os
 from PyQt4.QtCore import *
 from PyQt4.QtGui import *
 from qgis.core import *
-from osgeo import gdal
 from .utils import load_ui
 from ..geopackage_utils import GeoPackageUtils
-from ..grid_tools import raster2grid
+from ..user_communication import UserCommunication
+from ..grid_tools import raster2grid, grid_has_empty_elev
+import subprocess
 
 
 uiDialog, qtBaseClass = load_ui('sampling_elev')
 
 
 class SamplingElevDialog(qtBaseClass, uiDialog):
+
+    RTYPE = ['Byte', 'Int16', 'UInt16', 'UInt32', 'Int32', 'Float32', 'Float64']
 
     def __init__(self, con, iface, lyrs, cell_size, parent=None):
         qtBaseClass.__init__(self)
@@ -34,6 +37,7 @@ class SamplingElevDialog(qtBaseClass, uiDialog):
         self.setupUi(self)
         self.gutils = GeoPackageUtils(con, iface)
         self.gpkg_path = self.gutils.get_gpkg_path()
+        self.uc = UserCommunication(iface, 'FLO-2D')
         self.populate_raster_cbo()
         self.populate_alg_cbo()
         self.src_nodata = -9999
@@ -60,9 +64,10 @@ class SamplingElevDialog(qtBaseClass, uiDialog):
         if not self.srcRasterCbo.findData(self.src):
             bname = os.path.basename(self.src)
             self.srcRasterCbo.addItem(bname, self.src)
-            self.srcRasterCbo.setCurrentInsex(len(self.srcRasterCbo)-1)
+            self.srcRasterCbo.setCurrentIndex(len(self.srcRasterCbo)-1)
 
     def populate_alg_cbo(self):
+        """Populate resample algorithm combobox"""
         met = {
             "near": "Nearest neighbour",
             "bilinear": "Bilinear",
@@ -81,64 +86,22 @@ class SamplingElevDialog(qtBaseClass, uiDialog):
             self.algCbo.addItem(met[m], m)
             self.algCbo.setCurrentIndex(0)
 
-    def get_worp_options(self):
-        self.get_worp_opts_data()
-        self.wo = gdal.WarpOptions(
-            options=[],
-            format='GTiff',
-            outputBounds=self.output_bounds,
-            outputBoundsSRS=self.out_srs,
-            xRes=self.cell_size,
-            yRes=self.cell_size,
-            targetAlignedPixels=False,
-            width=0,
-            height=0,
-            srcSRS=self.src_srs,
-            dstSRS=self.out_srs,
-            srcAlpha=False,
-            dstAlpha=False,
-            warpOptions=None,
-            errorThreshold=None,
-            warpMemoryLimit=None,
-            creationOptions=['COMPRESS=LZW'],
-            outputType=self.raster_type,
-            workingType=self.raster_type,
-            resampleAlg=self.algCbo.itemData(self.algCbo.currentIndex()),
-            srcNodata=self.src_nodata,
-            dstNodata=self.src_nodata,
-            multithread=self.multiThreadChBox.isChecked(),
-            tps=False,
-            rpc=False,
-            geoloc=False,
-            polynomialOrder=None,
-            transformerOptions=None,
-            cutlineDSName=self.gpkg_path,
-            cutlineLayer=None,
-            cutlineWhere=None,
-            cutlineSQL="SELECT * FROM user_model_boundary",
-            cutlineBlend=None,
-            cropToCutline=False,
-            copyMetadata=True,
-            metadataConflictValue=None,
-            setColorInterpretation=False,
-            callback=None,
-            callback_data=None
-        )
-
     def get_worp_opts_data(self):
+        """Get all data needed for GDAL Warp"""
         # grid extents
         self.grid = self.lyrs.get_layer_by_name('Grid', self.lyrs.group).layer()
+        self.lyrs.update_layer_extents(self.grid)
         grid_ext = self.grid.extent()
         xmin = grid_ext.xMinimum()
         xmax = grid_ext.xMaximum()
         ymin = grid_ext.yMinimum()
         ymax = grid_ext.yMaximum()
-        self.output_bounds = (xmin, xmax, ymin, ymax)
+        self.output_bounds = (xmin, ymin, xmax, ymax)
         # CRS
         self.out_srs = self.grid.dataProvider().crs().toProj4()
         # data type
         src_raster_lyr = QgsRasterLayer(self.src_raster)
-        self.raster_type = src_raster_lyr.dataProvider().dataType(0)
+        self.raster_type = src_raster_lyr.dataProvider().dataType(1)
         self.src_srs = src_raster_lyr.dataProvider().crs().toProj4()
         # NODATA
         und = self.srcNoDataEdit.text()
@@ -146,15 +109,76 @@ class SamplingElevDialog(qtBaseClass, uiDialog):
             self.src_nodata = int(und)
 
     def probe_elevation(self):
-        """Resampling raster aligned with the grid"""
+        """Resample raster to be aligned with the grid, then probe values and update elements elevation attr"""
         self.src_raster = self.srcRasterCbo.itemData(self.srcRasterCbo.currentIndex())
         self.out_raster = '{}_interp.tif'.format(self.src_raster[:-4])
         try:
-            os.remove(self.out_raster)
+            if os.path.isfile(self.out_raster):
+                os.remove(self.out_raster)
         except OSError:
+            msg = 'Couldn\'t remove existing raster:\n{}\nChoose another filename.'.format(self.out_raster)
+            self.uc.show_warn(msg)
+            return False
+        self.get_worp_opts_data()
+        opts = [
+            '-of GTiff',
+            '-ot {}'.format(self.RTYPE[self.raster_type]),
+            '-tr {0} {0}'.format(self.cell_size),
+            '-te {}'.format(' '.join([str(c) for c in self.output_bounds])),
+            '-te_srs "{}"'.format(self.out_srs),
+            '-s_srs "{}"'.format(self.src_srs),
+            '-dstnodata {}'.format(self.src_nodata),
+            '-r {}'.format(self.algCbo.itemData(self.algCbo.currentIndex())),
+            '-co COMPRESS=LZW',
+            '-wo OPTIMIZE_SIZE=TRUE'
+        ]
+        if self.multiThreadChBox.isChecked():
+            opts.append('-multi -wo NUM_THREADS=ALL_CPUS')
+        else:
             pass
-        self.get_worp_options()
-        sampler = raster2grid(self.grid, self.out_raster, self.src_raster, self.wo)
+        cmd = 'gdalwarp {} {} {}'.format(' '.join([opt for opt in opts]), self.src_raster, self.out_raster)
+        proc = subprocess.Popen(
+            cmd,
+            shell=True,
+            stdout=subprocess.PIPE,
+            stdin=open(os.devnull),
+            stderr=subprocess.STDOUT,
+            universal_newlines=True,
+        ).stdout
+        for line in proc:
+            self.uc.log_info(line)
+        # Fill NODATA raster cells if desired
+        if self.fillNoDataChBox.isChecked():
+            self.fill_nodata()
+        else:
+            pass
+        sampler = raster2grid(self.grid, self.out_raster)
         qry = 'UPDATE grid SET elevation=? WHERE fid=?;'
         self.con.executemany(qry, sampler)
         self.con.commit()
+        return True
+
+    def fill_nodata(self):
+        opts = [
+            '-md {}'.format(self.radiusSBox.value())
+        ]
+        cmd = 'gdal_fillnodata {} {}'.format(' '.join([opt for opt in opts]), self.out_raster)
+        proc = subprocess.Popen(
+            cmd,
+            shell=True,
+            stdout=subprocess.PIPE,
+            stdin=open(os.devnull),
+            stderr=subprocess.STDOUT,
+            universal_newlines=True,
+        ).stdout
+        for line in proc:
+            self.uc.log_info(line)
+
+    def show_probing_result_info(self):
+        null_nr = grid_has_empty_elev(self.gutils)
+        if null_nr:
+            msg ='Sampling done.\n'
+            msg += 'Warning: There are {} grid elements that have no elevation value.'.format(null_nr)
+            self.uc.show_info(msg)
+        else:
+            self.uc.show_info("Sampling done.")
